# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalScene API - local discovery and booking platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: number;
  title: string;
  /** @nullable */
  body?: string | null;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  authorName: string;
  /** @nullable */
  authorAvatar: string | null;
  likes: number;
  commentsCount: number;
  createdAt: string;
  location: string;
  isBookable: boolean;
  liked: boolean;
  saved?: boolean;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  eventDate?: string | null;
  isFeatured?: boolean;
  isTrending?: boolean;
  /** @nullable */
  avgRating?: number | null;
  reviewsCount?: number;
}

export interface PostInput {
  title: string;
  body?: string;
  category: string;
  imageUrl?: string;
  location: string;
  isBookable: boolean;
  price?: number;
  eventDate?: string;
}

export interface Booking {
  id: number;
  postId: number;
  postTitle: string;
  /** @nullable */
  postImageUrl?: string | null;
  location: string;
  eventDate: string;
  status: string;
  /** @nullable */
  price?: number | null;
  createdAt: string;
  category?: string;
}

export interface BookingInput {
  postId: number;
}

export interface Profile {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  followersCount: number;
  followingCount: number;
  postsCount: number;
  /** @nullable */
  location?: string | null;
}

export interface ProfileUpdate {
  username?: string;
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  coverUrl?: string;
  location?: string;
}

export interface CalendarEvent {
  id: number;
  title: string;
  date: string;
  type: string;
  location: string;
  category?: string;
  /** @nullable */
  bookingId?: number | null;
  /** @nullable */
  price?: number | null;
}

export interface Comment {
  id: number;
  postId: number;
  authorName: string;
  body: string;
  createdAt: string;
}

export interface CommentInput {
  body: string;
  authorName?: string;
}

export interface Review {
  id: number;
  postId: number;
  authorName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  body: string;
  createdAt: string;
}

export interface ReviewInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  body: string;
  authorName?: string;
}

export interface SavedPlace {
  id: number;
  postId: number;
  createdAt: string;
  post: Post;
}

export interface SavePlaceInput {
  postId: number;
}

export interface PaymentMethod {
  id: number;
  /** @nullable */
  cardLabel?: string | null;
  cardLast4: string;
  cardBrand: string;
  expiryMonth: number;
  expiryYear: number;
  isDefault: boolean;
  createdAt: string;
}

export interface PaymentMethodInput {
  cardLabel?: string;
  cardLast4: string;
  cardBrand: string;
  expiryMonth: number;
  expiryYear: number;
}

export type StatsCategoryCounts = {[key: string]: number};

export interface Stats {
  totalPosts: number;
  upcomingBookings: number;
  totalBookings: number;
  categoryCounts: StatsCategoryCounts;
}

export type ListPostsParams = {
category?: string;
limit?: number;
};

export type UnsavePlace200 = {
  ok?: boolean;
};

export type DeletePaymentMethod200 = {
  ok?: boolean;
};

export type GetCalendarEventsParams = {
from?: string;
to?: string;
};