# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalScene API - local discovery and booking platform
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentMethod {
  id: number;
  /** @nullable */
  cardLabel?: string | null;
  cardLast4: string;
  cardBrand: string;
  expiryMonth: number;
  expiryYear: number;
  isDefault: boolean;
  createdAt: Date;
}