# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalScene API - local discovery and booking platform
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: number;
  title: string;
  /** @nullable */
  body?: string | null;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  authorName: string;
  /** @nullable */
  authorAvatar: string | null;
  likes: number;
  commentsCount: number;
  createdAt: Date;
  location: string;
  isBookable: boolean;
  liked: boolean;
  saved?: boolean;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  eventDate?: Date | null;
  isFeatured?: boolean;
  isTrending?: boolean;
  /** @nullable */
  avgRating?: number | null;
  reviewsCount?: number;
}