# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalScene API - local discovery and booking platform
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  postId: number;
  authorName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  body: string;
  createdAt: Date;
}