# lib/api-zod/src/generated/types/savePlaceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalScene API - local discovery and booking platform
 * OpenAPI spec version: 0.1.0
 */

export interface SavePlaceInput {
  postId: number;
}