# lib/api-zod/src/generated/types/savedPlace.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalScene API - local discovery and booking platform
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from './post';

export interface SavedPlace {
  id: number;
  postId: number;
  createdAt: Date;
  post: Post;
}